+++ expenses/db.go
@@ -1,0 +1,38 @@
+package expenses
+
+import (
+	"database/sql"
+	"fmt"
+	"log"
+	"os"
+
+	_ "github.com/lib/pq"
+)
+
+var db *sql.DB
+
+func InitDB() {
+	url := os.Getenv("DATABASE_URL")
+	var err error
+	db, err = sql.Open("postgres", url)
+	if err != nil {
+		log.Fatal("Connect to database error", err)
+	}
+
+	createTb := `
+	CREATE TABLE IF NOT EXISTS expenses (
+		id SERIAL PRIMARY KEY,
+		title TEXT,
+		amount FLOAT,
+		note TEXT,
+		tags TEXT[]
+	);
+	`
+	_, err = db.Exec(createTb)
+
+	if err != nil {
+		log.Fatal("can't create table", err)
+	}
+
+	fmt.Println("create table success")
+}
