+++ expenses/expenses.go
@@ -1,0 +1,13 @@
+package expenses
+
+type Expenses struct {
+	ID     int      `json:"id"`
+	Title  string   `json:"title"`
+	Amount float64  `json:"amount"`
+	Note   string   `json:"note"`
+	Tags   []string `json:"tags"`
+}
+
+type Err struct {
+	Message string `json:"message"`
+}
